MVP - Content
Added videos, images and short descriptions to services.

Added different videos for different pages.

Reduced the price of the Louby Lou doll for the website's launch!

To Do:
Convert images to WEBP for faster load times.
Format text in services to be more cohesive.
Alter services.json file to allow for a list of things to expect from a show.